Implement "small" memset()s via a _local_ memset and an RDMA put.
Because of the latencies of an AM, smaller memsets are faster via
RDMA puts.  However, we avoid doing this for large regions because the
bandwidth consumed on the wire is wasteful.
The "break even" point, in the absence of other traffic, appears to lie
between 32k and 128k,
However, to avoid consuming excessive bandwidth, the cut off is presently
the size of one bounce buffer (4k).  For regions larger than this the
AM-based memset is still used.