Remove unused thread argument and the corresponding macro wrappers.
The implementation of all the blocking gets and puts (but not memset)
are done directly, without any need to allocate a handle.  Therefore
they don't need any thread-specific data.

For this reason, the GASNETE_THREAD_FARG has been removed from
  gasnete_get_bulk()
  gasnete_put_bulk()

Similarly all the blocking calls which reduce to these two functions
no longer need to be implemented as macros which add the thread agrument
and call underscore-prepended inlines.  The following calls are now
implemented as inlines replacing the underscore-prepended inline versions:
  gasnet_get() gasnet_get_bulk() gasnet_get_val()
  gasnet_put() gasnet_put_bulk() gasnet_put_val()

Note that gasnet_memset() could be implemented directly as well, but I'll
do that later.