Fix polling routines to avoid deadlock.
The routines for polling RDMA completion have been restructured and
have become inline functions.

Previously a blocking RDMA operation would go into a tight spinloop
that checked only the send queue, and did not service any incomming
AMs.  That is fine when we have an extra thread that wakes on receive.
However, the spinloops now do a full AMPoll, ensuring that we could
drop the receive thread to yield a pure polling implementation which
is deadlock free.