Generalize/extend changes to gasnet_extended.h
I've been thinking about generalizing my changes to gasnet_extended.h so
that they can be merged witht he main branch.  This is what I've come up with.

I propose that this gasnet_extended.h replace the one in the main branch
when the vapi-conduit is merged.  I don't know of any reason why it should
break anything, but those are famous last words.
						-Paul

Present situation has some extended API calls as "direct" and others
as "derived".  The "direct" ones are those with gasnete_* variants
declared as externs in gasnet_extended.h, which the conduit must implement
if anything is to work.  The derived ones are those defined in terms of
one or more of the direct ones.  The present situation is this:

	put_nb			DIRECT
	put_nb_bulk		DIRECT
	get_nb			-> get_nb_bulk
	get_nb_bulk		DIRECT
	memset_nb		DIRECT
	try_syncnb		DIRECT
	try_syncnb_some		DIRECT
	try_syncnb_all		DIRECT
	wait_syncnb		-> try_syncnb
	wait_syncnb_some	-> try_syncnb_some
	wait_syncnb_all		-> try_syncnb_all

	put_nbi			DIRECT
	put_nbi_bulk		DIRECT
	get_nbi			-> get_nbi_bulk
	get_nbi_bulk		DIRECT
	memset_nbi		DIRECT
	try_syncnbi_gets	DIRECT
	try_syncnbi_puts	DIRECT
	try_syncnbi_all		DIRECT
	wait_syncnbi_gets	-> try_syncnbi_gets
	wait_syncnbi_puts	-> try_syncnbi_puts
	wait_syncnbi_all	-> try_syncnbi_gets + try_syncnbi_puts

	begin_nbi_accessregion	DIRECT
	end_nbi_accessregion	DIRECT

	put			-> put_bulk
	put_bulk		DIRECT / -> put_nb_bulk + wait_syncnb
	get			-> get_bulk
	get_bulk		DIRECT / -> get_nb_bulk + wait_syncnb
	memset			DIRECT / -> memset_nb + wait_syncnb

	put_val			-> put_nb + wait_syncnb
	put_nb_val		-> put_nb
	put_nbi_val		-> put_nbi
	get_val			-> get_nb_bulk + wait_syncnb
	get_nbi_val		DIRECT
	wait_syncnb_valget	DIRECT


Presently only three of the five blocking ops can be replaced by conduit-
specific direct implementations, and they must all be replaced as a group.
I want the ability for the conduit to replace any of the derived
implementations with their own direct versions.  In particular some of the
optimizations suggested by the spec are not possible unless one can
override some of your defaults implementations.  For instance if the
non-bulk gets are to optimize for small aligned transfers, they must not
always reduce to the *bulk* versions.

I want to make all of the functions derived above controllable with a
family of preprocessor tokens.  For example:

#if GASNETI_DIRECT_GET_BULK
  extern void gasnete_get_bulk (void *dest, gasnet_node_t node, void *src,
                                size_t nbytes GASNETE_THREAD_FARG);
#else
  GASNET_INLINE_MODIFIER(gasnete_get_bulk)
  void gasnete_get_bulk (void *dest, gasnet_node_t node, void *src,
                         size_t nbytes GASNETE_THREAD_FARG) {
    gasnete_wait_syncnb(gasnete_get_nb_bulk(dest, node, src,
                                            nbytes GASNETE_THREAD_PASS));
  }
#endif

I also want to change two of the default derivations, so that they
follow more natural paths of derivation.  For instance, if I have a
direct implementation of a non-bulk, blocking get which is optimized
for small aligned arguments (as the spec suggests) they the default
blocking get_val should use it, rather than the non-blocking bulk get
followed by a sync.  If I have no direct non-bulk, blocking get then
the actual implementation would reduce (transitivley) to the present
one.
	put_val		OLD -> put_nb + wait_syncnb		NEW -> put
	get_val		OLD -> get_nb_bulk + wait_syncnb	NEW -> get

I have implemented everything described above in vapi-conduit, where I
now have the following lines in gasnet_extended_fwd.h to reflect the
six operations I am currently overriding:

#define GASNETI_DIRECT_GET_BULK 1
#define GASNETI_DIRECT_PUT_BULK 1
#define GASNETI_DIRECT_MEMSET 1
#define GASNETI_DIRECT_WAIT_SYNCNB 1
#define GASNETI_DIRECT_WAIT_SYNCNBI_GETS 1
#define GASNETI_DIRECT_WAIT_SYNCNBI_PUTS 1