Don't set the fence bit on gets.
Setting the fence bit on RDMA READ ops prevents pipelining them.
However, it does not change the semantics any, since they are done
purely in order regardless.

This accounts for the big difference in GET performance between the
raw VAPI numbers taken for UPC benchmarking, and the ones I've been
measuring for GASNet.
				Gap	EEL
  vapi-conduit w/ fence bit	9.09us	9.02us
  vapi-conduit w/o fence bit	4.94us	6.27us
  Raw VAPI			4.95us	5.80us

I still want to investigate why the EEL is 0.47us higher than raw VAPI.
We only add about 0.28us to the EEL of the PUTS.