Implement first pass at robust gasnetc_exit()
Now "passes" all 7 cases in testexit, in that none of them hang.
However, #5 and #7 currently rely on the bootstrap to clean up orphans.

#5: "non-collective gasnet_exit(5), others in spin-loop"
    Currently the AM rcv thread is disabled to work around a VAPI bug.
    When this is restored, this test will exit gracefully.
#7: "non-collective gasnet_exit(7) between init()/attach()"
    Currently using AMs for graceful shutdown, but AM handlers are not
    registered until gasnet_attach().  This test will exit gracefully
    once I bypass AM to send the exit notifications.

There may also be some way to use the connection-oriented nature of
VAPI to "signal" threads when their peers close the connection.  That
would help reduce the chance of orphans.

I am leaving bug #120 open until I fix the issues w/ cases #5 and #7.