Cleanup exit-time behavior of snd/rcv engine.
Don't try to free srd/rcv resources in the single-node case, because
we didn't allocate any in this case!

There is a race in which the work queues might get destroyed out from
under us.  Therefore, when posting to the queues check for the appropriate
return code rather than failing an assertion and abort()ing in the middle
of a coordinated exit, thus fouling the exit code.

There are probably several more small races like this and they should
get treated similarlly as they are encountered.