Fix exit-time behavior of bootstrap code.
Testing shows that with LAM/MPI calling MPI_Finalize will ensure that the
mpirun returns an exit code of zedro, regardless of what the MPI tasks
actually _exit() with!

Therefore, I've make the bootstrapFini() a no-op for now.  This is not
elegant, but neither is use of MPI for bootstrapping GASNet, UPC, Ti.