Simplify AM flow control slightly and document a better solution.
While writing up my note on AM flow control for the README, I realized why
I still saw the RNR NAK asserted in some cases.  The reason is the window
of time between when a request handler sends a reply (thus granting a
credit) and the time when the replacement rbuf (the credit-counted
resource) is actually posted.

I had been going to some trouble to eliminate this window when we had to
send an "implicit" reply (because the handler sent none).  However, it is
not practical to eliminate the same window from the "explict" reply case.
Therefore, I've simplified the flow control code by not even trying to
avoid this small window in the implicit-reply case.

I've also documented (in the README) the "proper" solution, which is to
post a "spare" rbuf to the endpoint _before_ we begin to process an AM
REQUEST, rather than reposting the used one afterwards.  There are still
some implementation questions to be resolved however.

Also commiting some additional tracing/stats to track how often we need
to spin for an sbuf or credit, vs how often we get them on the first
try.  The related timings still lump them together for now.