Continue to refine the concurrency control in vapi-conduit.
Implemented two preprocessor tokens which describe, at compile time, the
type of concurrency which will be possible at runtime:
 GASNETC_ANY_PAR     Non-zero if multiple threads can be executing in GASNet.
                     This is inclusive of the AM receive thread.
 GASNETC_CLI_PAR     Non-zero if multiple _client_ threads can be executing in GASNet.
                     This excludes the AM receive thread.
See below for uses.


Replace GASNETC_MUTEX_* and GASNETC_PARLOCK_* with a different abstraction.
I've added a ultra-thin wrapper over gasneti_mutex_* which adds an extra
argument which determines (at compile time) if the lock op stays or goes:
  #define gasnetc_mutex_lock(X,C)              if (C) { gasneti_mutex_lock(X); }
  #define gasnetc_mutex_unlock(X,C)            if (C) { gasneti_mutex_unlock(X); }
  #define gasnetc_mutex_assertlocked(X,C)      if (C) { gasneti_mutex_assertlocked(X); }
  #define gasnetc_mutex_assertunlocked(X,C)    if (C) { gasneti_mutex_assertlocked(X); }
Example usage is
  gasnetc_mutex_lock(&schlage, GASNETC_CLI_PAR)


Add an abstract busy-waiting semaphore with just 4 ops:
  gasnetc_sema_init(&sema, n)
  gasnetc_sema_destroy(&sema)
  gasnetc_sema_up(&sema)
  gasnetc_sema_trydown(&sema, concurrency /* as with gasnetc_mutex_t */ )
Note this is a "busy-waiting" semaphore, in that there is only a "trydown"
rather than a "down" which would block.  I currently use this to deal with
counting of available transport resources.  When a trydown fails, we start
polling the network to reap the needed resources.

-Paul