Use the new gasnetc_sema_t to control depth of each send queue.
This is the start of the changes needed to make the resource usage more
scalable and configurable, and the associated code more maintainable.
Complete details of the resource management will go in the README when
it is all finallized.

Previously we allocated a fixed number (1024) of send descriptors,
independent of the number of nodes and then allowed them to be posted
to any endpoint.  The side effect of that is that each endpoint must
have a send queue large enough to enqueue ALL the send descriptors.
Under this scheme, if we started scaling the number of send descriptors
linearly with the number of nodes, the per-node cost of the send queues
would grow quadratically.

Now we can limit the number of sends posted to each send queue at runtime,
and will spin until space becomes available if we would exceed the limit.
Experimentation shows that a send queue of depth 16 is sufficient to "peg"
both small-message latency and large-message throughput for put_nbi_bulk.
The actual value is not so important as the fact that it is much less than
the 1024 I had before.

Now we can grow the number of allocated send resources linearly with a
corresponding linear growth in send queue resources.  Furthermore, we can
impose a maximum:
  send_descriptor_count = MIN(MAX_VALUE, 16*gasnetc_nodes)
The MAX_VALUE and 16 are both configurable, not hardcoded.  The names will
be documented in gasnet_core_internal.h and the README once I pick names
more intuitive than GASNETC_SND_CQ_SIZE and GASNETC_SND_WQE.