Significant changes/improvements to AM flow control in vapi-conduit.
The #1 change is that flow control on AMs is now required, rather than
being optional (could have been compiled away with the proper configuration
in gasnet_core_internal.h).  This change comes because I finally found the
example I need to prove that vapi-conduit could deadlock without it.

Two additional improvements are made as well:

Eliminated a race (not hard to trigger with the right testcase) in which
we would accept a credit from our peer (allowing us to send a AMRequest
from another thread) before we had a corresponding buffer posted for an
AMReply.

Use a little more intelligence in dealing with "reposting" of AM receive
buffers so that we don't crash if there is not a "spare" buffer available.
Instead we can carefully delay the granting of a credit until the running
handler completes, releasing its buffer for our use.  This results in a
longer latency for the granting of credits, but is only needed when we
are (temporarily) running low on resources.  This change involved putting
a "credit" bit into the flags word of the AM header.  This bit was stolen
from the 16-bit sourceID, and doesn't present a real problem since IB only
has a 14-bit LID space (we now also have one bit documented as unused).
With the use of an explicit credit bit (rather than assuming all AM replies
grant a credit) we open up the possibility of piggybacking credits on AM
requests as well.  This could reduce the latency of granting credits, for
instance, in the case where the main thread is making a AMRequest at the
same time we are handling one in the dedicated AM recv thread.  One example
is the roundtrip flood tests in gasnet/tests/testam.c.  I'll probably only
look into this once we have a better set of atomic ops so we can manage the
"outgoing" credits without needing a mutex.