Generalize the AM flow control mechanism for system-category AMs.
Now system-category AMs can request-reply-reply-reply without deadlock.
We don't actually have any such cases yet, but firehose might need this.

Also fixed a corner case bug in the previous flow control commit.
It turns out that the scheme used could still deadlock when resources were
scarce.  The alternative scheme now in use is to allow IB flow control
to kick in, but only for a bounded period of time (the time it takes for
an AM request handler to execute from its reply to the end) and only VERY
rarely.  It took a contrived case with artificially low resources to
cause this case to occur.