firehose checkin on new branch FIREHOSE_001
Checkin of a revised (not final) firehose spec.

There's also some initial implementation work which is just an expansion of
some of the functions in the spec.  The goal is to have a general interface
which constitutes the largest intersection possible between -region and -page
code.

Along with firehose.h, the following files can be found. .
	firehose_fwd.h - A sample forward file which will be moved to conduit
	                 directories once the implementation is ported;
	firehose_internal.h - An initial attempt at providing a private type
	                      for both -page and -region;
	firehose.c - An initial pseudo-code implementation.

   . . chris