firehose checkin: fixed some smp race conditions
Aside from race fixes, modified the code that runs firehose_callback so that no
locks are held while running the callback.  This was not a problem until it was
apparent that that some networks (such as GM) can only issue callbacks while
holding a global lock.  Now that firehose and conduits are separate, it's
probably only safe to call client callbacks without holding any locks.

Add a better mechanism for detecting remote pages that are currently in a
pending state (ones for which a firehose request is in progress and the reply
has not yet arrived).  At the cost of some extra time spent in the firehose
reply handler handling pending buckets, requests that hit pending buckets can
result in completion callbacks being delayed locally instead of requiring a
network roundtrip (an optimization useful in successive non-blocking operations
or threads targetting the same pages on the remote node).