firehose checkin: GM2 compatibility and firehose fixups.
This is the first checkin for complete (and tested) GM2 compatibility.  RDMA
get support looks promising with firehose.  I have some initial performance
graphs up at
http://mantis.lbl.gov/twiki/bin/view/User/GASNetGMConduit

Next I'll be running a few application benchmarks to see the effect of RDMA
gets (such as MG which makes extensive use of gets).

Unless some snags show up in the following days, I should be merging
GM2+firehose in the main branch soon.

FIREHOSE changes:

I've added an 'inhandler' argument to both firehose_release() and the firehose
callback type.  As the callback can be called from within or out of an AM
handler, it is useful (and sometimes required) for the client to know if the
call is in an AM handler.  Conversely, the client should inform firehose if it
is calling firehose_release() from within an AM handler.