Step 2 in the reorganization of firehose.
This commit removes all code and data related to the bucket tables and
the type fh_bucket_t from the common code.  The only remaining reference
to bucket_t (other than a bunch of comments that need updating) is a
forward decl of the type "struct _fh_bucket_t", needed to keep a pointer
to one in a single place.

From this point forward "firehose_private_t" is the "common" link between
-page and -region.  While the common code is responsible for the FIFOs,
all lookups, creation and destruction of private_t takes place in the
files firehose_page.c and firehose_region.c.