Next stage of firehose reorganization.
Move the abstraction boundary which separates firehose.c from
firehose-{page,region}.c.  Now instead of using region_t and lots of
copying in and out, the functions such as fh_commit_try_local_region()
and fh_acquire_remote_region() work on the request_t directly.
This also provides the path needed for firehose_region.c to set the
"internal" field of a request_t.

This now places all calls to fh_request_{new,free}() in firehose.c,
allowing them to become static (and perhaps later inline as well).

At this point vapi-conduit is able to process firehose "hits" against
the prepinned segments at a cost of <1000 cycles for
  p = firehose_try_local_pin();
  firehose_release(&p, 1);

Only firehose_release() and firehose_{try,partial}_{local,remote}_pin() are
fully implemented at this point, and the remote versions can't succeed until
we have a full implementation of firehose_remote_pin().

Next step is to implement firehose_local_pin() so that we can pin stuff
on local firehose "misses".
After that comes the remote stuff which is going to involve trying to
generalize the AM stuff.