Fully implement fh_{acquire,release}_local_region() in firehose-region.c.
With this commit, I've implemented the complete set of local pinning
functions:
  firehose_local_pin()
  firehose_try_local_pin()
  firehose_partial_local_pin()
  firehose_release() with only local regions.
Testing w/ a hacked up vapi-conduit confirms that the generalized code for
the FIFO is working correctly.  I can see a mixture of firehose hits and
misses, and FIFO evictions.  It also appears that the hash which picks the
"best" match is probably working, but this is not absolutely verified yet.