Fix badly-performing region merging in firehose-region.
The performance-killing problem noted in a previous commit was due to
actually unpinning any regions which were merged into a new request.
Instead we now just allow the normal victim FIFO to ensure eventual
reuse of the region "obscured" by a newly created region.

Also tweaked the hash table handling a little so that when two regions
covering a given page have the same ending address, the one with the
larger total length will be chosen.  Otherwise, the ending address is
used for comparison (as it was before) under a sort of "read ahead" type
of assumption.