initial shmem core conduit implementation done.
Here's an untested, although complete shmem core.  As is, this should work on
the altix and the X1 -- since the shmem interface on the altix is almost
identical to the X1, and since configure runs 1000 times faster than the X1,
I'll be testing the conduit on the SGI machine first (... it's even has lower
latency than the X1).

All in all, this adds about 500-600 lines to template-conduit.  It took me 1h
to fix configure, and about 4 hours to implement the core.  Although the
conduit is not tested, it is based on a core algorithm that I've refined and
previously tested over various platforms.  In fact, the code that's checked in
has portions selectively optimized for Cray shmem, SGI shmem and quadrics
shmem.

Absentees from this checkin are. .

* I winged gasnet segment support.  Currently have a binary-search applied
  to shmalloc (similar to what we have with mmap).  I expect the gasnet segment
  to be the most troublesome part of shmem-conduit as each vendor does this
  differently.  It's easy on Cray and relatively easy on SGI. . quadrics needs
  the elan allocator.

* Spent some time doing some configure hacking (detection, linking, etc).
  Still pending are things like disabling -par configuration, since shmem is
  inherintly not threadsafe.