sgi/shmem core works
Few fixes for sgi/shmem.  All gasnet tests run on sgi shmem, and there's some
detection in configure to try to disambiguate what flavour of shmem is found on
a given Linux box.

Left is porting to elan/shmem which requires some tinkering with the elan
allocator.  Then the extended api.