gexec spawner modifications
Yet another round of fixups for a new setup over at the Millennium.  I've added
environment variable propagation in the script, and will follow up to this in
the conduit if time permits.

Environment variables GASNET_GEXEC_CMD and GASNET_GSTAT_CMD can be set to
control the operation of the spawner with gexec.