shmem-conduit extended API
This checks in a near-complete shmem-conduit gasnet.  Right now, I've only been
able to test it on the Altix (since the X1 is down) but most of it should just
work on the X1.  What's left to do. .

* Bug in AM-based gasnet_memset that I'm aware of

* The gasnet barrier still uses the AM barrier.  It will be difficult to get a
  good barrier over shmem, because of notify/wait and the fact that AMPoll must
  be run at a barrier.  This approach blows off shmem_barrier_all() as a
  possible candidate.  Here are our options. .
    * Use the shmem_test_event (only available on Cray shmem)
    * Spawn an other thread to call shmem_int_wait (which is on every shmem).
      This may not be possible.
    * Some gather/broadcast or tree-based/pairwise exchange with RDMA puts
      (Paul, were you working on this?)

* More aggressive inlining.  Right now a good deal of functions can be inlined
  but there are others which could benefit from direct shmem calls (which would
  allow our compiler to spit out shmem directly).

* Gets out of segment may not work on all shmems.  For one, getting into a
  stack variable is problematic on some shmems.  I'd expect the value gets to
  be of use for stack variables but even non-value gets must support in-stack
  gets.

We pass all gasnet tests except testthreads (assumed) and the memset portion of
testgasnet.

We need to turn off the huge caches on these Itaniums for sensible
testsmall/large numbers or else we get 100ns latency and 12Gb/s bandwidth.

Dan: please review changes to gasnet_extended.h.  The approach is similar as
for the 'direct' implementation of some gasnet functions, although this time it
is for inlines (these inlines are in gasnet_extended_fwd.h).