Start a branch for implementing one of the ideas in bug #267
The idea of chaining multiple operations for reducing the number of
completion records is a potentially beneficial one, but it has the
possibility of causing deadlock since the resources can't be reused until
a completion record is generated.  The mechanisms to avoid this possible
deadlock may make the code complicated enough to eliminate the potential
benefits.

This is an experiment I'll be playing with over the holiday break.  So,
I am doing it in a branch in case it fails to produce any benefit.