More remote firehose work for firehose-region branch. The remote pinning now works correctly: 1) The region merging is done on the responder end where it can be done    much more agressively. 2) The correct table entries are now created on the responder (fixing a    bug in the previous major commit).
Note that pinning operations (both local and in the AM handler) currently
lack the desired thread safety.  In particular, the table entries are not
created until AFTER the reacquisition of the lock following the pin.  The
way firehose-region is built, this just creates a harmless duplicate (or
subset) pin in the case of a race.  Unpinning is safe because regions are
distinct in that a given page remains pinned until *all* pinned regions
which contain it have been unpinned.  However, it is desirable to eliminate
this race because it leads to wastefull use of resources.

Next steps:
1) factor some code common to local and remote cases.
2) thread safetly - create a PENDING entry for *local* memory before we
   release the lock to invoke the move callback.
3) vapi-conduit work to implement the extended API over firehose.