Back out yesterday's thread-safety work in firehose-region.
I realized on my way home that the solution I had adopted was hopelessly
broken in the presence of interrupt-driven AM implementations.  Rather
then just throwing in a no-interrupts region around the call to the
conduit's move_callback, I have decided to look at other options first.

For now there is only one "problem" with multiple threads (including the
AM rcv thread in vapi-conduit) in which duplicate local pinnings of
identical regions can result.  At the moment that breaks the hash which
is used to match the unpin portion of a move AM to the correct table
entry.  It is also potentially quite wastefull since in the case of
multiple indentical regions produced by AMs handled concurrently, there is
a very real chance that none of the duplicates will ever get released.

Proper use of hold/resume interrupt seems critical to the correct
operation of both region- and page-based firehose.  This needs further
discussion w/ Dan and Christian.