Implement GASNET_SEGMENT_EVERYTHING (and LARGE) in vapi-conduit! :-)
This is (for now) still in a separate branch.

This is about the simplest implementation possible of the extended API in
terms of the firehose interface.  Effort has been made to optimize only
the cost of a firehose *hit*.  The results are encouraging:

Comparing EVERYTHING to FAST (in which the entire segment is pre-pinned)
the end-to-end latency (EEL) and initiator-side overhead (o_i) of small
gets and puts appear to have increased by between 0.4 and 0.5us.  The o_i
number is still smaller than the gap, and thus peak throughput for the
non blocking operations is indistinguishable between the segment models
(execept that for transfers > 128kB the EVERYTHING results are actually
*better* than FAST - but this due to a known issue with how the HCA is
caching virtual->physical translations).

Work remains on the vapi-conduit/extended-api end to reduce the cost of
a firehose miss.  In particular the following might be possible:
1) Use the "remote callback" in firehose to piggyback a PUT on the move AM.
2) Use the "remote callback" in firehose to piggyback a GET on the move AM.
   This assumes the initiator end is already pinned.
3) Overlap the remote and local pins.
4) Use the firehose_*_partial_pin() functions to interleave RDMA to parts
   of the source/destination which are already pinned with the move AM(s)
   required to pin the remainder.

Work still remains on the firehose end to get proper thread safety.