Implement the first optimization to vapi-conduit's use of firehose. We now try to overlap local firehose operations with the possible remote miss.
Doing so results in measureable improvement when we miss at both ends.
Before this change missing at both ends on a PUT resulted in an effective
EEL of roughly 197us.  With the overlap the effective EEL drops to 131us.
For comparison, the cost of a remote miss with local hit is 125us.  Thus
with the overlap we pay only an added 6us for missing at the local end in
addition to the remote miss.