Update region merging heuristic based on recent experiments.
Given a sequence of touches which include all pages in the segment (and
assuming they can all be pinned at once), the current algorithm creates an
"efficient" covering of the region in the sense that "live" regions in the
hash do not overlap (i.e. no page appears in more than one live region).
Here a live region is defined as one which may receive future hits.  The
smaller regions created as firehose-region learns the accesible addresses
are not live.

It is able to do this while also making the regions as large as reasonably
possible within the compile-time bound.  The result is that for in-order
or reverse-order touches of all pages one gets a "perfect" coverage which
includes only maximum-lengh regions (possibly excepting the first or last
region).  For random-order touches the average number of regions needed
apears (empirically) to be about 1.5 times the number needed for the
"perfect" coverage.

All previous alogrithms used here had pathalogical behavior for at least
one of the forward, backward or random ordering.