Add a testconduit for firehose-smp
This testconduit is really a toy conduit, with just enough functionality to
test the firehose code.  The conduit will start up 'P' processes and 'T'
threads, and communicate AMs using local TCP sockets.  The idea is to provide a
mechanism by which the firehose code (smp and uni variants) can be stress
tested for race conditions.

I think the conduit may also be useful for benchmarking firehose usage, such as
gathering put/get GASNet traces from real applications and re-running the
traces through the testconduit.  For example, we could report:

   * Firehose hit ratio for application benchmarks
   * Effectiveness of sharing the firehose state table across processors when
     using GASNet PAR (which is the whole point of this project).