Start firehose-page-smp integration
Integrate most of the pseudocode from local and remote into firehose_page and
decide on a policy to switch between uni and smp versions of firehose-page --
they will both remain within the same .c file and implement a uni and an smp
version of each internal firehose call that requires thread-safety.

Also, theres a few modifications for the testconduit, although it's not ready
for prime time yet.