Fix performance bug when large number of nb broadcasts were posted.
Improve the general framework to be easier to read and to allow more
compile-time optimization (such as inlining) by replacing use of function
pointers in one place with a switch.  This also puts related code together.

TO DO: There is still a need for a "big switch" at the top level.
This can be done w/o sacraficing conduit-specific customizations using a
mechanism similar to what Dan has done for the tracing infrastructure.