Updates for many fixes and smaller issues we didn't catch
Too many to enumerate here, but most of the updates are either conditions we
hadn't thought about -- parts of the code made assumptions about never seeing a
pending state which caused problems in various areas.

(the real reason for this commit is that alvarez is acting up and I don't want
to lose my work!).