Implement IN_MYSYNC (bcast only) using point-to-point synchronizations. This removes the need for a global barrier at the start of the collective.
For small payload we use an "eager" mechanism to send the data to be
buffered until the target enters the collective.

For large payload (or for Titanium-type clients with only local address
information) we send the source addr to be buffered at each destination.
When the target node enters the collective it will Get the specified data
from the origin.

Note that has only been done for the case of aligned segments w/o pthreads.
I'll next be doing the general case (includes pthreads and/or unaligned
segments in a single implementation).

Also fixed two pretty serious bugs in the point-to-point synchronization
infrastructure I commited earlier today.