Fix bug 839 - missing UPC file/line in gasnet trace
Handling the high-level line numbers correctly can be rather complicated (trace calls from header inline functions being one major source of grief), so sometimes a tweak or two is required.

I've just committed a fix for the problem, and a fix to a different but related issue, that caused all pthread/debug backend jobs to crash at or near startup if tracing was enabled.

I'm committing this to the release branch only because upc_trace is a
high priority for 2.0.1 and the second bug caused tracing to be fatally broken
on all pthreads configurations. I'm 95% certain these fixes won't break anything
but it would be very nice to do a branch nightly test on this release candidate -
do we have that capability yet?