Start work on custom VIS implemenation for vapi-conduit. This is intended as a small project for me to work on on the SC04 floor. I hope this branch will be short-lived, with a merge at the end of SC04.
This initial commit applies the NIC's gather-on-read capability to the
use of bounce buffers for Gets with out-of-segment destinations.  Using
this capability, we can post 4 bounce buffers, of 4k each, as a single vapi
operation.  Compared to posting single bounce buffers, this increases the
peak bandwidth of gasnet_get_nb_bulk from roughly 580000 Kb/s to roughly
740000 Kb/s (a better than 27% increase).
As this makes the out-of-segment Get bandwidth higher than the out-of-segment
Put bandwidth, the next logical step is to apply is to that problem as well.