Start a branch (TREE_COLLECTIVES), in which I am merging Rajesh's work with the reference GASNet collectives.  This is taking place in a branch because there is some non-trivial rewritting to do (such as generalizing the code to use as a base for teams support) that may result in code that is broken in significant ways at various times.
This first commit just begins adding the data structure and its init/fini code.

Note that this branch tag is applied only to the extended-ref directory.
The (non-branch) tag TREE_COLLECTIVES_mergepoint will be used to mark the
"corresponding" point on the main branch (to be moved as any changes made
to the main branch are merged into this branch).