Initial (naive) implementation of dynamically pinned bounce buffers.
This implementation still allocates the maximum amount of memory, but doesn't
pre-pin it.  This implementation does not honor GASNET_USE_FIREHOSE=no in
the sense that firehose is used for the bounce buffers regardless.
This is at least partially broken for SEGMENT_{LARGE,EVERYTHING} since to
send a firehose-move AM requires allocating a bounce buffer, which is to
be dynamically pinned locally, by firehose.  A sufficient number of threads
could end up making so many requests that things might deadlock.  I need to
think more on that point, however, because this might be resolved simply by
having correct resource limits.
CVSMAIL: CC: