temporary fix for tree free problem
the problem with freeing trees was caused by a duplicate call to tree put. It was temporarily patched by removing the possibiity of a duplicate free. However upon further discussions with Dan, we think that the true solution lies in implementing a lazy tree creation scheme and a mechanism that never frees the trees once they are created.

Thus once a new tree geometry is requested it will be created and added to a linked list of existing trees. When we ask for that same geometry again, the implementation will go through the link list and fetch this new tree data structure. This is the next implementation step on this branch. Full details will be available in a forthcoming design document.
CVSMAIL: CC: