added tree based all-gather and many major fixes
fixed barrier implementations that required folding the barrier
(can't go down the tree then up the tree ... *duh*)

fixed scratch managment to work with all-gather (very minor changes)

there is still a problem with the barrier allocation with-in the scratch space management. The barriers are outside the collective framework. Still need to fix it. The main thing this currently precludes is the use of non-blocking collectives. Next checkin will addresss this problem
CVSMAIL: CC: