added new all_gather implementation
i just added a ring based all-gather routine. It will take (P-1) steps to finish the communication. I don't expect it to do that well for small messages but it might be helpful in large messages. MPICH implements this routine so i figure that it should be one of the routines that we have in our suite as well.


CVSMAIL: CC: