added segmented versions of scatter and gather
these now require an extra argument to be passed into the poll function that specifies the distance between the portions for different threads. Before it was trivially the transfer size, howveer if we only do a portion of the collective we need to specify how far apart they are.

Unfortuneatly this also prohibits us from taking advantage of certain optimizations to avoid copying to and from scratch space. If our distance between successive portions to different processors is greater than the transfer size then the data is no longer contiguous and thus we need to use the extra buffer space to make the transfers contiguous.

The algorithms as written right now also do not relfect our thoughts on how to properly deal with the aggregation flags and what flags to pass into the various collectives. This issue should be an easy fix once we agree on what needs to be done.
CVSMAIL: CC: