Fix an error in gasnet debug malloc tracing.
Was unconditionally testing to see if a pointer was in the shared
region, but if check happens before gasnet_attach is complete, then we
get a segmentation fault (because gasneti_seginfo is NULL).

Fixed by checking for gasneti_attach_done != 0.  Dan: let me know if
this isn't the right thing to check (should we check gasneti_seginfo !=
NULL instead?).

CVSMAIL: CC: