Minor change to sysvnet API.
Change gasnet_sysvnet_t to typedef the struct itself, rather than a
struct to a pointer to the struct.

This just keeps things in line with the rest of the typedefs in the
implementation.  The struct is still opaque to clients, so an error will
be emitted if one tries to directly declare (or access members of) the
struct.

CVSMAIL: CC: