Make gasneti_segmentAttach() SysV-friendly.
If the original mmapped region was too big, code used to munmap the
whole thing and then re-map the subregion.  This doesn't work for
SysV--you can't re-mmap the shared region after we've forked--so I've
replaced the full-remap logic with code to only munmap the part of the
original region that needs to be trimmed.

Unless there's a bug in my code (me? right bugs?), this new code should
work fine for both the SysV and non-SysV cases, but for now I'm keeping
it in an #ifdef.   Gasnet Poobahs and peanut gallery members are
invited to weigh in on whether the #ifdef should be removed.

CVSMAIL: CC: